Implement `clippy` command.
This commit implements a `clippy` command that uses `clippy-driver` as the
`rustc` wrapper.

This command effectively behaves as `cargo clippy`, but respects the generated
bindings crates.

We'll remove this command once upstream clippy is patched such that it respects
the `CARGO` environment variable for running `cargo check` (allowing it to use
`cargo-component check` to do the actual check).

Closes #13.