Update dependencies and latest component model spec.
This commit updates the dependencies and, consequently, pulls in a
`wit-component` that uses the latest component model spec. It also changes the
wit-bindgen dependencies back to upstream now that changes needed to implement
`cargo component` have been merged.

Minor changes to the build and check commands for adding a `--keep-going`
option to match latest cargo.

Updated the generated `wit` file for `cargo component new` to the latest wit
syntax, as well as a few test files.