Update `Cargo.toml` format for `cargo component`.
This commit updates the format for `Cargo.toml` to make it clearer as to what
gets imported and what gets exported.

It splits the former `dependencies` table into two tables `imports` and
`exports`.

Each dependency in those tables is uniformly specified as either:

```toml
dep = "path"
```

or:

```toml
dep = { path = "path" }
```

This removes the version numbers from the dependencies for now until such time
as they are useful.

A warning was added if we encounter the old `dependencies` table to warn users
they need to upgrade their `Cargo.toml` format; the tool doesn't attempt to do
so automatically (sorry).

Closes #28.