Update `Cargo.toml` format for `cargo component`.  (#29)
* Use more string interpolation in `cargo-component`.

This commit updates some string and error message formatting to use string
interpolation / named parameters vs. positional parameters.

* Update `Cargo.toml` format for `cargo component`.

This commit updates the format for `Cargo.toml` to make it clearer as to what
gets imported and what gets exported.

It splits the former `dependencies` table into two tables `imports` and
`exports`.

Each dependency in those tables is uniformly specified as either:

```toml
dep = "path"
```

or:

```toml
dep = { path = "path" }
```

This removes the version numbers from the dependencies for now until such time
as they are useful.

A warning was added if we encounter the old `dependencies` table to warn users
they need to upgrade their `Cargo.toml` format; the tool doesn't attempt to do
so automatically (sorry).

Closes #28.

* Rename `default` option to `direct-interface-export`.

This commit renames the `default` option in `Cargo.toml` to
`direct-interface-export` to make it explicitly clear as to what the option is
for.