Remove opinionated package identifier format.
This commit removes the previous expectation that a package identifier is
formatted a certain way (previously `<namespace>/<name>`).

The warg protocol is likely to not require a particular format for a package
identifier string, rather individual registries may enforce a format for
their own namespacing of packages.

This ultimately makes `PackageId` function like a string, the meaning of which
is up to the registry being communicated with.