Implement a `--target` option for `cargo component new`.
This commit implements a `--target` option that causes `cargo component new` to
set the target to use in `Cargo.toml`.

It also generates `src/lib.rs` based on the exports of the target work,
providing the necessary boilerplate code for users to immediately start
implementing the component's exports.

A `--registry` option is also added to set the default registry of the project
and is currently used to resolve the `--target` option until an internal
default registry is implemented.

Additionally, a `--world` option was added to set the world to use when
specifying a target WIT package.