Merge the `document` and `world` settings.
This commit merges the `document` and `world` settings in Cargo.toml for
specifying the target world to use from a package.

Additionally, it updates the resolution to first use `Resolve::select_world` to
find the world to use.

As binary-encoded WIT packages don't currently encode a notion of a "default"
world, the previous fallback is kept for parsing binary WIT packages.

Closes #51.