Allow WIT package dependencies to be shared.
This commit removes calls to `Resolve::push_dir` when processing target
dependencies.

As a result, no WIT files will be parsed from a `deps` directory automatically,
so each dependency must be specified in the target dependencies table.

This change performs roughly the same work that `push_dir` was doing, namely
the topological sort of the dependencies so that they are parsed in the correct
order.

It should now be possible to share a target dependency between other target
dependencies, much like one would expect from the `deps` directory behavior
from `Resolve::push_dir`.