Code review feedback.
Move the code generator and the use trie to a separate module from the bindings
module.

Fix the use trie to emit children and types in lexographical order; also fix
where the emitting of braces so that they are only written if there is more
than one child and/or type written.