Add an optional configuration to disable `rustfmt`
Currently `wit-bindgen` is a relatively dumb code generator and the
generated code can fail in `rustfmt` due to long lines and things like
that. I'm not sure how best to fix this in `wit-bindgen` or how best to
communicate this to `rustfmt` (or if it's even a bug there). One option
would be to ignore `rustfmt` failures in `wit-bindgen` but that seems
not great because then there's not a great means to communicate the
non-fatal error.

In lieu of solving some of these thornier issues I figured it'd be good
to at least have an escape hatch in `cargo component` to disable
`rustfmt` in case errors crop up again in the future.