Fix toposort for decoded binary WIT packages.
Previously, the toposort performed for WIT dependencies would skip visiting any
of the dependencies of decoded binary packages because it was assumed that it
was unnecessary due to the fact the dependency information was encoded in the
package.

However, wit-parser doesn't merge interfaces unless they have identical
definitions (see https://github.com/bytecodealliance/wasm-tools/issues/1191).

This means that we run into problems when a decoded binary package is merged
*before* a dependency that contains the full definition of the interface.

To work around this issue, users should be able to add an explicit dependency
on any transitive dependencies so that they get merged first.

For that to work, we need to visit any dependencies on decoded binary packages
as well, which is what this commit implements.