Update WASI adapters to 2ad057d
This is for compatibility with the Wasmtime 13 release branch.

I have retained the previous adapters (134dddc) because I believe the
plan is to support switching between them soon, and I imagine there
would still be people wanting to use these older adapters even if
running the latest version of cargo-component.