Add an `adapter` setting to `Cargo.toml`.
This commit adds a `package.metadata.component.adapter` field to `Cargo.toml`
which is the path to the WASI module adapter to use.

If present, the specified file is used to adapt WASI preview1 to preview2.

If not present, the built-in WASI module adapter is used instead.

The intention with this commit is to allow users control over how WASI is
adapted in cases where the built-in adapter is not the desired one to use.