Encode the module before the adapter in componentization.
This commit changes the order in which we componentize the core module by
calling `adapter` after `module`.

This matches what we do in `wasm-tools component new` and also makes it more
obvious when there is a mismatch between any WASI definitions local to the
module and the preview 1 adapter, as it will favor the local definitions first.