Fix bindings generation for exports of versioned interfaces.
This commit fixes bindings generation for the exports of interfaces coming from
versioned packages (i.e. the WIT package directive has a version on it).

The issue is that the export map passed through to wit-bindgen is expecting
_not_ to have the version on the interface name as the version is associated
with the package and not the interface.

The fix is to stop appending the version number when populating the exports
map.