Refactor bindings generation.
This commit alters how bindings generation works in `cargo-component`.

Previously, `cargo-component` would encode a target world for the `generate!`
macro to decode. The macro would decode the target and feed it to
`wit-bindgen` and expand to the generated bindings.

This approach had a few drawbacks:

* compilation times for projects increased as `wit-bindgen` had to be built
  from the component itself.
* rust-analyzer would not re-expand the `generate!` macro on WIT file changes as
  it does not add a source dependency via `include_bytes!`.
* "go to definition" doesn't work with rust-analyzer as the bindings code was
  entirely from the macro expansion.

To fix all three issues, `cargo-component` now generates the bindings source
file itself and the macro simply expands to `include!` it.

This cuts the initial crate dependencies for a project from 50 to 13, speeding
up the build (and our CI).

And as the `include_bytes!` is now an `include!`, rust-analyzer adds a source
file dependency to re-expand the macro when any local WIT files change.

However, this comes at the cost of being a breaking change:

* if the `cargo-component-bindings` crate version referenced from the project
  is mismatched from `cargo-component` following this change, the component
  project will not build as they won't agree on how to generate bindings.
* any users that provide arguments to the previous `generate!` macro must move
  those arguments into the `[package.metadata.component.bindings]` section in
  `Cargo.toml`.

Fixes #139.