Fix bindings version warning and upgrade command.
This commit fixes the warning emitted when a mismatched bindings version crate
is detected such that it no longer warns when the bindings crate version is
compatible, but not the same version (i.e. `0.4.1` when cargo-component is
`0.4.0`).

Likewise, the `upgrade` command no longer attempts to downgrade the version of
the bindings crate when it would otherwise be compatible.

The upgrade tests no longer attempt the install, which require a long time to
just verify that a `cargo install cargo-component` would succeed, which isn't
terribly useful.