Improve the UX for supporting the run/test commands.
This commit pipes the stdout and inherits stderr when spawning `cargo`,
allowing for immediate output from cargo without buffering.

It also improves the processing of wasm artifacts to componentize; the
"creating component" message no longer prints on every build for macOS users.

Removed forcing colors in cargo output as we now inherit stderr, allowing cargo
to detect if it is a terminal or not.