Remove `cargo_component_bindings` crate.
This commit is a breaking change to `cargo-component` with the intention of
improving component build time (now down to two transitive crate dependencies)
and get `rust-analyzer` fully working with bindings.

It removes the `cargo_component_bindings` and `cargo_component_macro` crates.
Instead of using a macro for including the generated bindings,
`cargo-component` now directly writes to a `src/bindings.rs` file which is
included as a submodule in `main.rs` or `lib.rs`.

Additionally, the `update` subcommand was removed as it is no longer necessary
since there is no `cargo_component_bindings` crate referenced in `Cargo.toml`;
the functionality to update `cargo-component` itself can simply be replaced
with a `cargo install` or `cargo binstall`.

Closes #210.