Implement `unlocked-dep` imports for component dependencies.
This commit implements outputting a component with `unlocked-dep` imports for
any component dependencies specified in `Cargo.toml`.

Note that WIT does not support the nesting of interfaces (i.e.
instances), so to properly bind to a component's exports, an instance is
imported for its directly exported functions and then one for every instance
the component exports.

Therefore, depending on a component may result in more than one import
dependency for that package; one of the imports might be for
`unlocked-dep=<foo:bar>`, while others might be of the form
`unlocked-dep=<foo:bar/<export>>`.