Fix panic in bindings generation for component dependencies. (#253)
This commit fixes a panic originating in `wit-bindgen` that occurs when a
component dependency exports an interface that contains a used type.

Previously, the bindings generation only added used interfaces for types
directly used by the component's world (i.e. from exported functions).

The fix was to do the same for any exported interfaces as well.

Fixes #252.