Synchronize artifact componentization.
This commit makes componentization of build artifacts acquire the lock file in
an effort to prevent concurrent cargo-component processes that are building the
same workspace from interfering with one another.

Additionally, this commit changes the writing of the final build artifact so
that it writes to a temp file and then atomically persists the temp file at the
final path.

Fixes #268.