Bug fixes for targeting a world with resource borrows (#300)
Expanded test cases to cover issue #299

Also, when specifying an exact semver requirement, adds that exact requirement to the generated Cargo.toml. There was mismatch before.

wit-bindgen-rust generates different bindings for borrow<> of resources, depending on whether that type is exported in the world or not. &SomeType vs SomeTypeBorrow in the generated bindings. Fixed the logic here to match that behavior. The revised test case should better cover future potential breakage.

If interested, here is some of the logic on the wit-bindgen side:
https://github.com/bytecodealliance/wit-bindgen/blob/main/crates/rust/src/interface.rs#L2203-L2225